Create keypressregisterwithtimestamps.py
This script is just a basic keylogger that records when keys are pressed by recording the keypress and the time. This was made to get some data to determine typing speed and help narrow in remaining latency from point of seeing something to better get an idea of the distance a server was behind a cloudflare gateway based on RTL between when an image was seen and when a key was pressed when served from the internet, subtracting the known traceroute, and keypress time. This was just one part that let me record the keypresses and time the response times.